+++ plugin.py
@@ -16,23 +16,13 @@ for module_name in [
 prefix = None
 
 from .unittesting.color_scheme import UnitTestingColorSchemeCommand
-from .unittesting.coverage import UnitTestingCoverageCommand
-from .unittesting.current import UnitTestingCurrentFileCommand
-from .unittesting.current import UnitTestingCurrentFileCoverageCommand
-from .unittesting.current import UnitTestingCurrentPackageCommand
-from .unittesting.current import UnitTestingCurrentPackageCoverageCommand
-from .unittesting.package import UnitTestingCommand
 from .unittesting.syntax import UnitTestingSyntaxCommand
 from .unittesting.syntax import UnitTestingSyntaxCompatibilityCommand
+from .unittesting.unit import UnitTestingCommand
 
 
 __all__ = [
     "UnitTestingCommand",
-    "UnitTestingCoverageCommand",
-    "UnitTestingCurrentFileCommand",
-    "UnitTestingCurrentFileCoverageCommand",
-    "UnitTestingCurrentPackageCommand",
-    "UnitTestingCurrentPackageCoverageCommand",
     "UnitTestingSyntaxCommand",
     "UnitTestingSyntaxCompatibilityCommand",
     "UnitTestingColorSchemeCommand",
@@ -44,34 +34,11 @@ __all__ = [
 sys.modules["unittesting"] = sys.modules["UnitTesting"].unittesting
 
 UT33_CODE = """
-from UnitTesting import plugin as ut38  # noqa
+from UnitTesting import plugin as ut38
 
 
 class UnitTesting33Command(ut38.UnitTestingCommand):
-    pass
-
-
-class UnitTesting33CoverageCommand(ut38.UnitTestingCoverageCommand):
-    pass
-
-
-class UnitTesting33CurrentPackageCommand(ut38.UnitTestingCurrentPackageCommand):
-    pass
-
-
-class UnitTesting33CurrentPackageCoverageCommand(ut38.UnitTestingCurrentPackageCoverageCommand):
-    pass
-
-
-class UnitTesting33CurrentFileCommand(ut38.UnitTestingCurrentFileCommand):
-    pass
-
-
-class UnitTesting33CurrentFileCoverageCommand(ut38.UnitTestingCurrentFileCoverageCommand):
-    pass
-
-
-class UnitTesting33ColorSchemeCommand(ut38.UnitTestingColorSchemeCommand):
+    \"\"\"Execute unit tests for python 3.3 plugins.\"\"\"
     pass
 """
