--- unittesting/const.py
@@ -1,1 +1,0 @@
-DONE_MESSAGE = "UnitTesting: Done.\n"
