+++ unittesting/scheduler.py
@@ -1,72 +1,81 @@
 import os
 import sublime
 
-from .utils import JsonFile
-
-
-class Unit:
-
-    def __init__(self, s):
-        self.package = s['package']
-
-        self.output = s.get('output', None)
-        self.syntax_test = s.get('syntax_test', False)
-        self.syntax_compatibility = s.get('syntax_compatibility', False)
-        self.color_scheme_test = s.get('color_scheme_test', False)
-        self.coverage = s.get('coverage', False)
-
-    def run(self):
-        if self.syntax_test:
-            sublime.run_command("unit_testing_syntax", {
-                "package": self.package,
-                "output": self.output
-            })
-        elif self.syntax_compatibility:
-            sublime.run_command("unit_testing_syntax_compatibility", {
-                "package": self.package,
-                "output": self.output
-            })
-        elif self.color_scheme_test:
-            sublime.run_command("unit_testing_color_scheme", {
-                "package": self.package,
-                "output": self.output
-            })
-        elif self.coverage:
-            sublime.run_command("unit_testing_coverage", {
-                "package": self.package,
-                "output": self.output,
-                "generate_xml_report": True
-            })
-        else:
-            sublime.run_command("unit_testing", {
-                "package": self.package,
-                "output": self.output
-            })
+__all__ = ["run_scheduler"]
 
 
 class Scheduler:
-
-    def __init__(self):
-        self.units = []
-        self.j = JsonFile(os.path.join(
-            sublime.packages_path(),
-            'User', 'UnitTesting',
-            'schedule.json')
+    def run(self):
+        schedule = Schedule(
+            os.path.join(
+                sublime.packages_path(), "User", "UnitTesting", "schedule.json"
+            )
         )
+        for s in schedule.load():
+            Unit(s).run()
+        schedule.save([])
 
-    def load_schedule(self):
-        self.schedule = self.j.load()
-        for s in self.schedule:
-            self.units.append(Unit(s))
+
+class Schedule:
+    def __init__(self, fpath, encoding="utf-8"):
+        self.encoding = encoding
+        self.fpath = fpath
+
+    def load(self, default=[]):
+        try:
+            with open(self.fpath, "r", encoding=self.encoding) as fp:
+                json_data = sublime.decode_value(fp.read())
+                if not isinstance(json_data, list):
+                    raise ValueError("Unsupported data format!")
+                return json_data
+        except FileNotFoundError:
+            return default
+        except Exception as e:
+            print("Error loading {}: {!s}".format(self.fpath, e))
+            return default
+
+    def save(self, data, indent=4):
+        self.fdir = os.path.dirname(self.fpath)
+        os.makedirs(self.fdir, exist_ok=True)
+        with open(self.fpath, "w", encoding=self.encoding) as fp:
+            fp.write(sublime.encode_value(data, True))
+
+
+class Unit:
+    def __init__(self, s):
+        self.package = s["package"]
+
+        self.output = s.get("output", None)
+        self.syntax_test = s.get("syntax_test", False)
+        self.syntax_compatibility = s.get("syntax_compatibility", False)
+        self.color_scheme_test = s.get("color_scheme_test", False)
+        self.coverage = s.get("coverage", False)
 
     def run(self):
-        self.load_schedule()
-        for u in self.units:
-            u.run()
-        self.clean_schedule()
-
-    def clean_schedule(self):
-        self.j.save([])
+        if self.color_scheme_test:
+            sublime.active_window().run_command(
+                "unit_testing_color_scheme",
+                {"package": self.package, "output": self.output},
+            )
+        elif self.syntax_test:
+            sublime.active_window().run_command(
+                "unit_testing_syntax", {"package": self.package, "output": self.output}
+            )
+        elif self.syntax_compatibility:
+            sublime.active_window().run_command(
+                "unit_testing_syntax_compatibility",
+                {"package": self.package, "output": self.output},
+            )
+        else:
+            sublime.active_window().run_command(
+                "unit_testing",
+                {
+                    "package": self.package,
+                    "output": self.output,
+                    "coverage": self.coverage,
+                    "generate_xml_report": True,
+                },
+            )
 
 
 def run_scheduler():
