+++ unittesting/utils/__init__.py
@@ -1,11 +1,9 @@
-from .json_file import JsonFile
 from .output_panel import OutputPanel
 from .progress_bar import ProgressBar
 from .stdio_splitter import StdioSplitter
 from .reloader import reload_package
 
 __all__ = [
-    "JsonFile",
     "OutputPanel",
     "StdioSplitter",
     "ProgressBar",
